+++ src/modules/esl-carousel/renderers/esl-carousel.default.renderer.ts
@@ -1,8 +1,7 @@
-import {promisifyTransition} from '../../esl-utils/async';
-import {normalize, normalizeIndex} from '../core/nav/esl-carousel.nav.utils';
+import {normalize, sign} from '../core/esl-carousel.utils';
 import {ESLCarouselRenderer} from '../core/esl-carousel.renderer';
 
-import type {ESLCarouselDirection} from '../core/nav/esl-carousel.nav.types';
+import type {ESLCarouselDirection, ESLCarouselActionParams} from '../core/esl-carousel.types';
 
 /**
  * Default carousel renderer based on CSS Flexbox stage, order (flex), and stage animated movement via CSS transform.
@@ -23,17 +22,27 @@ export class ESLDefaultCarouselRenderer extends ESLCarouselRenderer {
 
   /** Slides gap size */
   protected gap: number = 0;
-  /** Slide size cached value */
-  protected slideSize: number = 0;
   /** First index of active slides. */
   protected currentIndex: number = 0;
+
+  /** Multiplier for the index move on the slide move */
+  protected get INDEX_MOVE_MULTIPLIER(): number {
+    return 1;
+  }
+
+  /** Actual slide size (uses average) */
+  protected get slideSize(): number {
+    return this.$slides.reduce((size, $slide) => {
+      return size + (this.vertical ? $slide.offsetHeight : $slide.offsetWidth);
+    }, 0) / this.$slides.length;
+  }
 
   /**
    * Processes binding of defined renderer to the carousel {@link ESLCarousel}.
    * Prepare to renderer animation.
    */
   public override onBind(): void {
-    this.currentIndex = this.$carousel.activeIndex >= 0 ? this.$carousel.activeIndex : 0;
+    this.currentIndex = this.normalizeIndex(Math.max(0, this.$carousel.activeIndex));
     this.redraw();
   }
 
@@ -62,36 +71,43 @@ export class ESLDefaultCarouselRenderer extends ESLCarouselRenderer {
     if (!slide) return 0;
     return this.vertical ? slide.offsetTop : slide.offsetLeft;
   }
+
   /** Sets scene offset */
   protected setTransformOffset(offset: number): void {
     this.$area.style.transform = `translate3d(${this.vertical ? `0px, ${offset}px` : `${offset}px, 0px`}, 0px)`;
   }
 
-  /** @returns current slide area's transform offset */
-  protected getTransformOffset(): number {
-    // computed value is matrix(a, b, c, d, tx, ty)
-    const transform = getComputedStyle(this.$area).transform;
-    if (!transform || transform === 'none') return 0;
-    const position = this.vertical ? 5 : 4; // tx or ty position
-    return parseInt(transform.split(',')[position], 10);
+  /** Animates scene offset to index */
+  protected async animateTo(index: number, duration = 250): Promise<void> {
+    this.currentIndex = this.normalizeIndex(index);
+    const offset = -this.getOffset(this.currentIndex);
+    this.$carousel.$$attr('animating', true);
+    await this.$area.animate({
+      transform: [`translate3d(${this.vertical ? `0px, ${offset}px` : `${offset}px, 0px`}, 0px)`]
+    }, {duration, easing: 'linear'}).finished;
+    this.$carousel.$$attr('animating', false);
   }
 
   /** Pre-processing animation action. */
-  public override async onBeforeAnimate(): Promise<void> {
+  public override async onBeforeAnimate(nextIndex: number, direction: ESLCarouselDirection, params: ESLCarouselActionParams): Promise<void> {
     if (this.$carousel.hasAttribute('animating')) throw new Error('[ESL] Carousel: already animating');
     this.$carousel.$$attr('active', true);
   }
 
   /** Processes animation. */
-  public async onAnimate(nextIndex: number, direction: ESLCarouselDirection): Promise<void> {
+  public async onAnimate(nextIndex: number, direction: ESLCarouselDirection, params: ESLCarouselActionParams): Promise<void> {
     const {activeIndex, $slidesArea} =  this.$carousel;
     this.currentIndex = activeIndex;
     if (!$slidesArea) return;
-    while (this.currentIndex !== nextIndex) await this.onStepAnimate(direction === 'next' ? 1 : -1);
+    const distance = normalize((nextIndex - activeIndex) * direction, this.size);
+    const speed = Math.min(1, this.count / distance);
+    while (this.currentIndex !== nextIndex) {
+      await this.onStepAnimate(direction * this.INDEX_MOVE_MULTIPLIER, params.stepDuration * speed);
+    }
   }
 
   /** Post-processing animation action. */
-  public override async onAfterAnimate(): Promise<void> {
+  public override async onAfterAnimate(nextIndex: number, direction: ESLCarouselDirection, params: ESLCarouselActionParams): Promise<void> {
     // Make sure we end up in a defined state on transition end
     this.reorder();
     this.setTransformOffset(-this.getOffset(this.currentIndex));
@@ -99,39 +115,24 @@ export class ESLDefaultCarouselRenderer extends ESLCarouselRenderer {
   }
 
   /** Makes pre-processing the transition animation of one slide. */
-  protected async onStepAnimate(indexOffset: number): Promise<void> {
+  protected async onStepAnimate(indexOffset: number, duration: number): Promise<void> {
     const index = normalize(this.currentIndex + indexOffset, this.size);
 
     // Make sure there is a slide in required direction
     this.reorder(indexOffset < 0);
-
     const offsetFrom = -this.getOffset(this.currentIndex);
     this.setTransformOffset(offsetFrom);
 
-    // here is the final reflow before transition
-    const offsetTo = -this.getOffset(index);
-    // Allow animation and move to the target slide
-    this.$carousel.$$attr('animating', true);
-    this.setTransformOffset(offsetTo);
-    if (offsetTo !== offsetFrom) {
-      await promisifyTransition(this.$area, 'transform');
-    }
-
-    this.currentIndex = index;
-    this.$carousel.$$attr('animating', false);
-  }
-
-  protected indexByOffset(count: number): number {
-    return this.$carousel.activeIndex + count;
+    await this.animateTo(index, duration);
   }
 
   /** Handles the slides transition. */
-  public onMove(offset: number): void {
+  public move(offset: number, from: number, params: ESLCarouselActionParams): void {
     this.$carousel.toggleAttribute('active', true);
 
     const slideSize = this.slideSize + this.gap;
     const count = Math.floor(Math.abs(offset) / slideSize);
-    const index = this.indexByOffset(count * (offset < 0 ? 1 : -1));
+    const index = from + count * this.INDEX_MOVE_MULTIPLIER * (offset < 0 ? 1 : -1);
 
     // check left border of non-loop state
     if (!this.loop && offset > 0 && index <= 0) return;
@@ -143,39 +144,37 @@ export class ESLDefaultCarouselRenderer extends ESLCarouselRenderer {
 
     const stageOffset = this.getOffset(this.currentIndex) - (offset % slideSize);
     this.setTransformOffset(-stageOffset);
+
+    if (this.currentIndex !== this.$carousel.activeIndex) {
+      this.setActive(this.currentIndex, {direction: sign(-offset)});
+    }
   }
 
   /** Ends current transition and make permanent all changes performed in the transition. */
   // eslint-disable-next-line sonarjs/cognitive-complexity
-  public async commit(offset: number): Promise<void> {
+  public async commit(offset: number, from: number, params: ESLCarouselActionParams): Promise<void> {
     const slideSize = this.slideSize + this.gap;
 
     const amount = Math.abs(offset) / slideSize;
     const tolerance = ESLDefaultCarouselRenderer.NEXT_SLIDE_TOLERANCE;
     const count = (amount - Math.floor(amount)) > tolerance ? Math.ceil(amount) : Math.floor(amount);
-    const index = this.indexByOffset(count * (offset < 0 ? 1 : -1));
+    const index = from + count * this.INDEX_MOVE_MULTIPLIER * (offset < 0 ? 1 : -1);
 
-    this.currentIndex = normalizeIndex(index, this);
-
-    // Hm ... that's what actually happens on slide step
-    this.$carousel.$$attr('animating', true);
-    const stageOffset = -this.getOffset(this.currentIndex);
-    this.setTransformOffset(stageOffset);
-    if (stageOffset !== this.getTransformOffset()) {
-      await promisifyTransition(this.$area, 'transform');
-    }
-    this.$carousel.$$attr('animating', false);
+    await this.animateTo(index);
 
     this.reorder();
     this.setTransformOffset(-this.getOffset(this.currentIndex));
-    this.setActive(this.currentIndex, {direction: offset < 0 ? 'next' : 'prev'});
     this.$carousel.$$attr('active', false);
+
+    if (this.currentIndex !== this.$carousel.activeIndex) {
+      this.setActive(this.currentIndex, {direction: sign(-offset)});
+    }
   }
 
   /**
    * @returns count of slides to be rendered (reserved) before the first slide
    */
-  protected calcReserveCount(back?: boolean): number {
+  protected getReserveCount(back?: boolean): number {
     const {size, count, loop, currentIndex} = this;
     const freeSlides = size - count;
     // no need to reorder if there are no free slides or loop is disabled
@@ -183,7 +182,7 @@ export class ESLDefaultCarouselRenderer extends ESLCarouselRenderer {
     // if back option is not set, prefer to reserve slides with respect to semantic order
     if (typeof back !== 'boolean') back = !!currentIndex;
     // otherwise, ensure that there are at least half of free slides reserved (if the back option is set - round up, otherwise - round down)
-    return Math.min(count, back ? Math.ceil(freeSlides / 2) : Math.floor(freeSlides / 2));
+    return back ? Math.ceil(freeSlides / 2) : Math.floor(freeSlides / 2);
   }
 
   /**
@@ -192,7 +191,7 @@ export class ESLDefaultCarouselRenderer extends ESLCarouselRenderer {
    */
   protected reorder(back?: boolean): void {
     const {size, loop, currentIndex, $slides} = this;
-    const reserve = this.calcReserveCount(back);
+    const reserve = this.getReserveCount(back);
 
     const index = loop ? currentIndex : 0;
     for (let i = 0; i < size; ++i) {
@@ -210,7 +209,7 @@ export class ESLDefaultCarouselRenderer extends ESLCarouselRenderer {
 
     this.gap = parseFloat(this.vertical ? areaStyles.rowGap : areaStyles.columnGap);
     const areaSize = parseFloat(this.vertical ? areaStyles.height : areaStyles.width);
-    this.slideSize = Math.floor((areaSize - this.gap * (this.count - 1)) / this.count);
-    this.$area.style.setProperty(ESLDefaultCarouselRenderer.SIZE_PROP, this.slideSize + 'px');
+    const slideSize = Math.floor((areaSize - this.gap * (this.count - 1)) / this.count);
+    this.$area.style.setProperty(ESLDefaultCarouselRenderer.SIZE_PROP, slideSize + 'px');
   }
 }