+++ src/modules/esl-carousel/renderers/esl-carousel.grid.renderer.ts
@@ -2,8 +2,7 @@ import {prop, memoize} from '../../esl-utils/decorators';
 import {ESLCarouselRenderer} from '../core/esl-carousel.renderer';
 import {ESLDefaultCarouselRenderer} from './esl-carousel.default.renderer';
 
-import type {ESLCarouselDirection} from '../core/nav/esl-carousel.nav.types';
-import type {ESLCarouselActionParams} from '../core/esl-carousel';
+import type {ESLCarouselActionParams} from '../core/esl-carousel.types';
 
 /**
  * {@link ESLDefaultCarouselRenderer} extension to render slides as a multi-row grid.
@@ -22,8 +21,14 @@ export class ESLGridCarouselRenderer extends ESLDefaultCarouselRenderer {
   @prop(2, {readonly: true})
   public readonly ROWS: number;
 
-  /** @returns count of fake slides to fill the last "row" or incomplete carousel state */
+  /** Multiplier for the index move on the slide move */
+  protected override get INDEX_MOVE_MULTIPLIER(): number {
+    return this.ROWS;
+  }
+
+  /** Count of fake slides to fill the last "row" or incomplete carousel state */
   public get fakeSlidesCount(): number {
+    if (this.$carousel.$slides.length === 0) return 0;
     if (this.$carousel.$slides.length < this.count) {
       return this.count - this.$carousel.$slides.length;
     }
@@ -41,7 +46,7 @@ export class ESLGridCarouselRenderer extends ESLDefaultCarouselRenderer {
     return Array.from({length}, this.buildFakeSlide.bind(this));
   }
 
-  /** @returns all slides including {@link ESLGridCarouselRenderer.$fakeSlides} slides created in grid mode */
+  /** All slides including {@link ESLGridCarouselRenderer.$fakeSlides} slides created in grid mode */
   public override get $slides(): HTMLElement[] {
     return (this.$carousel.$slides || []).concat(this.$fakeSlides);
   }
@@ -73,31 +78,17 @@ export class ESLGridCarouselRenderer extends ESLDefaultCarouselRenderer {
   }
 
   /**
-   * Processes changing slides
    * Normalize actual active index to the first slide in the current dimension ('row')
    */
-  public override async navigate(index: number, direction: ESLCarouselDirection, {activator}: ESLCarouselActionParams): Promise<void> {
-    await super.navigate(index - (index % this.ROWS), direction, {activator});
-  }
-
-  /** Processes animation. */
-  public override async onAnimate(nextIndex: number, direction: ESLCarouselDirection): Promise<void> {
-    const {activeIndex, $slidesArea} =  this.$carousel;
-    this.currentIndex = activeIndex;
-    if (!$slidesArea) return;
-    const step = this.ROWS * (direction === 'next' ? 1 : -1);
-    while (this.currentIndex !== nextIndex) await this.onStepAnimate(step);
-  }
-
-  protected override indexByOffset(offset: number): number {
-    return super.indexByOffset(offset * this.ROWS);
+  protected override normalizeIndex(index: number, params?: ESLCarouselActionParams): number {
+    return super.normalizeIndex(index - (index % this.ROWS), params);
   }
 
   /**
    * @returns count of slides to be rendered (reserved) before the first slide does not include fake slides
    */
-  protected override calcReserveCount(back?: boolean): number {
-    const reserve = super.calcReserveCount(back);
+  protected override getReserveCount(back?: boolean): number {
+    const reserve = super.getReserveCount(back);
     return reserve - (reserve % this.ROWS);
   }
 
@@ -109,7 +100,7 @@ export class ESLGridCarouselRenderer extends ESLDefaultCarouselRenderer {
     this.gap = parseFloat(this.vertical ? areaStyles.rowGap : areaStyles.columnGap);
     const areaSize = parseFloat(this.vertical ? areaStyles.height : areaStyles.width);
     const count = Math.floor(this.count / this.ROWS);
-    this.slideSize = Math.floor((areaSize - this.gap * (count - 1)) / count);
-    this.$area.style.setProperty(ESLDefaultCarouselRenderer.SIZE_PROP, this.slideSize + 'px');
+    const slideSize = Math.floor((areaSize - this.gap * (count - 1)) / count);
+    this.$area.style.setProperty(ESLDefaultCarouselRenderer.SIZE_PROP, slideSize + 'px');
   }
 }