+++ src/modules/esl-carousel/renderers/esl-carousel.none.renderer.ts
@@ -1,7 +1,7 @@
 import {ESLCarouselRenderer} from '../core/esl-carousel.renderer';
 
 import type {ESLCarousel} from '../core/esl-carousel';
-import type {ESLCarouselConfig, ESLCarouselDirection} from '../core/nav/esl-carousel.nav.types';
+import type {ESLCarouselActionParams, ESLCarouselConfig, ESLCarouselDirection} from '../core/esl-carousel.types';
 
 /**
  * None effect carousel renderer. Does not provide any animation, transition. Does not limit slide stage.
@@ -29,20 +29,12 @@ export class ESLNoneCarouselRenderer extends ESLCarouselRenderer {
     this.setActive(0);
   }
 
-  public override onUnbind(): void {
-    // this.$carousel.scrollTop = this.$carousel.scrollLeft = 0;
-  }
+  public override onUnbind(): void {}
 
   /** Processes animation. */
-  public async onAnimate(nextIndex: number, direction: ESLCarouselDirection): Promise<void> {
-  }
+  public async onAnimate(nextIndex: number, direction: ESLCarouselDirection): Promise<void> {}
 
-  /** Handles the slides transition. */
-  public onMove(offset: number): void {
-    // TODO: implement if scroll behaviour requested
-    // const property = this.vertical ? 'scrollTop' : 'scrollLeft';
-    // this.$carousel[property] = -offset;
-  }
-  public commit(offset?: number): void {
-  }
+  /* Handles the slide move actions */
+  public move(offset: number, from: number, params: ESLCarouselActionParams): void {}
+  public async commit(offset: number, from: number, params: ESLCarouselActionParams): Promise<void> {}
 }